# swagger.yaml
openapi: 3.0.3
info:
  title: CrusherMate API
  description: |
    **CrusherMate Backend API** - Complete truck entry management system for crusher material business.

    ## Features
    - User authentication and authorization
    - Truck entry management (Sales & Raw Stone)
    - Material rate management
    - Dashboard analytics
    - Financial reporting
    - Real-time data export
    - Organization management

    ## Authentication
    All protected endpoints require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Response Format
    All API responses follow a consistent format:
    ```json
    {
      "success": true,
      "message": "Response message",
      "data": { /* response data */ },
      "timestamp": "2024-01-01T00:00:00.000Z"
    }
    ```

    ## Error Handling
    Error responses include appropriate HTTP status codes and error details:
    ```json
    {
      "success": false,
      "message": "Error message",
      "error": "Detailed error information",
      "timestamp": "2024-01-01T00:00:00.000Z"
    }
    ```

  version: "1.0.0"
  contact:
    name: CrusherMate Support
    email: support@crushermate.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://your-api-domain.vercel.app
    description: Production server
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        totalDocs:
          type: integer
          description: Total number of documents
        limit:
          type: integer
          description: Number of documents per page
        totalPages:
          type: integer
          description: Total number of pages
        page:
          type: integer
          description: Current page number
        hasNextPage:
          type: boolean
          description: Whether there is a next page
        hasPrevPage:
          type: boolean
          description: Whether there is a previous page

    # User Related Schemas
...
    Organization:
      type: object
      properties:
        id:
          type: string
        _id:
          type: string
        name:
          type: string
          maxLength: 100
        owner:
          type: string
          description: Owner user ID
        members:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Truck Entry Schemas
    TruckEntry:
      type: object
      properties:
        id:
          type: string
        _id:
          type: string
        organization:
          type: string
          description: Organization ID or name
        userId:
          type: object
          properties:
            _id:
              type: string
            username:
              type: string
        truckNumber:
          type: string
          pattern: "^[A-Z]{2}[0-9]{2}[A-Z]{1,2}[0-9]{4}$"
          description: Truck number (e.g., KA01AB1234)
        truckName:
          type: string
          maxLength: 50
          description: Driver name or truck identifier
        entryType:
          type: string
          enum: [Sales, Raw Stone]
          description: Type of entry
        materialType:
          type: string
          enum:
            [
              M-Sand,
              P-Sand,
              Blue Metal 0.5in,
              Blue Metal 0.75in,
              Jally,
              Kurunai,
              Mixed,
            ]
          description: Material type (required for Sales entries)
        units:
          type: number
          minimum: 0.1
          maximum: 100
          description: Number of units
        ratePerUnit:
          type: number
          minimum: 1
          description: Rate per unit in INR
        totalAmount:
          type: number
          description: Auto-calculated total amount
        truckImage:
          type: string
          description: URL to truck image
        entryDate:
          type: string
          format: date
          description: Date of entry
        entryTime:
          type: string
          pattern: "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"
          description: Time of entry (HH:MM format)
        status:
          type: string
          enum: [active, deleted]
          default: active
        notes:
          type: string
          maxLength: 500
          description: Additional notes
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TruckEntryCreate:
      type: object
      properties:
        truckNumber:
          type: string
          pattern: "^[A-Z]{2}[0-9]{2}[A-Z]{1,2}[0-9]{4}$"
          description: Truck number (e.g., KA01AB1234)
        truckName:
          type: string
          maxLength: 50
          description: Driver name or truck identifier
        entryType:
          type: string
          enum: [Sales, Raw Stone]
          description: Type of entry
        materialType:
          type: string
          enum:
            [
              M-Sand,
              P-Sand,
              Blue Metal 0.5in,
              Blue Metal 0.75in,
              Jally,
              Kurunai,
              Mixed,
            ]
          description: Material type (required for Sales entries)
        units:
          type: number
          minimum: 0.1
          maximum: 100
          description: Number of units
        ratePerUnit:
          type: number
          minimum: 1
          description: Rate per unit in INR
        entryDate:
          type: string
          format: date
          description: Date of entry
        entryTime:
          type: string
          pattern: "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"
          description: Time of entry (HH:MM format)
        notes:
          type: string
          maxLength: 500
          description: Additional notes
      required:
        - truckNumber
        - truckName
        - entryType
        - units
        - ratePerUnit
        - entryDate
      example:
        truckNumber: "KA01AB1234"
        truckName: "Ravi Kumar"
        entryType: "Sales"
        materialType: "M-Sand"
        units: 10
        ratePerUnit: 2200
        entryDate: "2024-01-15"
        entryTime: "14:30"
        notes: "Quality material"

    # Material Rate Schema
    MaterialRate:
      type: object
      properties:
        id:
          type: string
        materialType:
          type: string
          description: Type of material
        currentRate:
          type: number
          description: Current rate per unit
        lastUpdated:
          type: string
          format: date-time
        updatedBy:
          type: string
          description: User who updated the rate
        organization:
          type: string
          description: Organization ID
        isActive:
          type: boolean
          default: true

    # Dashboard Summary Schema
...
    OtherExpense:
      type: object
      properties:
        id:
          type: string
        organization:
          type: string
          description: Organization ID
        user:
          type: string
          description: User ID who created the expense
        expensesName:
          type: string
          maxLength: 100
          description: Type of expense
        amount:
          type: number
          minimum: 1
          description: Expense amount
        others:
          type: string
          maxLength: 500
          description: Additional details
        date:
          type: string
          format: date
          description: Date of expense
        isActive:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Report Schema
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Detailed error information
        timestamp:
          type: string
          format: date-time
      required:
        - success
        - message
        - timestamp
paths:
...
  /api/truck-entries:
    get:
      tags:
        - Truck Entries
      summary: Get truck entries
      description: Retrieve truck entries with pagination and filtering
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
            default: entryDate
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: truckNumber
          in: query
          description: Filter by truck number
          schema:
            type: string
        - name: entryType
          in: query
          description: Filter by entry type
          schema:
            type: string
            enum: [Sales, Raw Stone]
        - name: materialType
          in: query
          description: Filter by material type
          schema:
            type: string
        - name: startDate
          in: query
          description: Start date for filtering
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: End date for filtering
          schema:
            type: string
            format: date
        - name: userId
          in: query
          description: Filter by user ID (owner only)
          schema:
            type: string
      responses:
        "200":
          description: Truck entries retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TruckEntry"
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "400":
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - Truck Entries
      summary: Create truck entry
      description: Create a new truck entry (Sales or Raw Stone)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: "#/components/schemas/TruckEntryCreate"
                - type: object
                  properties:
                    truckImage:
                      type: string
                      format: binary
                      description: Truck image file
          application/json:
            schema:
              $ref: "#/components/schemas/TruckEntryCreate"
      responses:
        "201":
          description: Truck entry created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Truck entry created successfully"
                  data:
                    type: object
                    properties:
                      truckEntry:
                        $ref: "#/components/schemas/TruckEntry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/material-rates:
    get:
      tags:
        - Material Rates
      summary: Get material rates
      description: Get all material rates for the organization
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Material rates retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MaterialRate"

    post:
      tags:
        - Material Rates
      summary: Update material rate (Owner only)
      description: Create or update a material rate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                materialType:
                  type: string
                  description: Type of material
                rate:
                  type: number
                  minimum: 1
                  description: Rate per unit
              required:
                - materialType
                - rate
              example:
                materialType: "M-Sand"
                rate: 2200
      responses:
        "201":
          description: Rate updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Rate for M-Sand updated successfully"
                  data:
                    $ref: "#/components/schemas/MaterialRate"
        "403":
          description: Owner access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Configuration Endpoints
...
  /api/expenses:
    get:
      tags:
        - Other Expenses
      summary: Get other expenses
      description: Get all other expenses with filtering and pagination
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Other expenses retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/OtherExpense"

    post:
      tags:
        - Other Expenses
      summary: Create other expense
      description: Create a new other expense entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                expensesName:
                  type: string
                  maxLength: 100
                amount:
                  type: number
                  minimum: 1
                others:
                  type: string
                  maxLength: 500
                date:
                  type: string
                  format: date
              required:
                - expensesName
                - amount
                - date
              example:
                expensesName: "Fuel"
                amount: 5000
                others: "Fuel for crushing machine"
                date: "2024-01-15"
      responses:
        "201":
          description: Other expense created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/OtherExpense"
...
  /api/organizations:
    get:
      tags:
        - Organizations
      summary: Get all organizations
      description: Get list of all organizations (admin only)
      responses:
        "200":
          description: Organizations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Organization"

    post:
      tags:
        - Organizations
      summary: Create organization
      description: Create a new organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 100
              required:
                - name
      responses:
        "201":
          description: Organization created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    $ref: "#/components/schemas/Organization"